fix: normalize entity models across all services
Key fixes:
- Double→BigDecimal for monetary fields (Producto, CatalogoItem)
- String→LocalDateTime for date fields (Pedidos.fechaPedido, Inventarios.fechaActualizacion)
- String→@Enumerated(EnumType.STRING) for state fields (Despacho.estado, Pedidos.estadoPedido)
- Integer→BigDecimal for DetallePedidos.precioUnitario
- Added @Column(name=...), @PrePersist/@PreUpdate, Bean Validation missing in multiple entities
- Added EstadoDespacho and EstadoPedido enums with type-safe state transitions
- Fixed Proveedor table name (proveedor→proveedores), added unique constraint on correo, audit fields
- Removed comment garbage in CatalogoCampana and Resena

+++ catalogo/src/main/java/duoc/cl/catalogo/dto/CatalogoItemDTO.java
@@ -2,6 +2,7 @@ package duoc.cl.catalogo.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @Schema(description = "DTO de un item dentro de una campaña del catálogo")
@@ -18,9 +19,9 @@ public class CatalogoItemDTO {
     @Schema(description = "Nombre del proveedor")
     private String nombreProveedor;
     @Schema(description = "Precio en catálogo")
-    private Double precioCatalogo;
+    private BigDecimal precioCatalogo;
     @Schema(description = "Precio de oferta")
-    private Double precioOferta;
+    private BigDecimal precioOferta;
     @Schema(description = "Estado del stock")
     private String estadoStock;
 }
