+++ catalogo/src/main/java/duoc/cl/catalogo/dto/CatalogoRequest.java
@@ -3,6 +3,7 @@ package duoc.cl.catalogo.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 import lombok.Data;
 
@@ -19,12 +20,12 @@ public class CatalogoRequest {
     @Positive(message = "precio debe ser mayor a cero")
     @NotNull(message = "el precio catalogo es obligatorio")
     @Schema(description = "Precio en catálogo", example = "19990")
-    private Double precioCatalogo;
+    private BigDecimal precioCatalogo;
 
     // precio oferta obligatorio
     @Positive(message = "precio debe ser mayor a cero")
     @NotNull(message = "el precio oferta es obligatorio")
     @Schema(description = "Precio de oferta", example = "15990")
-    private Double precioOferta;
+    private BigDecimal precioOferta;
 
 }
