+++ pedidos/src/main/java/com/duoc/pedidos/dto/DetallePedidoDTO.java
@@ -3,11 +3,13 @@ package com.duoc.pedidos.dto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Schema(description = "DTO de un detalle de pedido")
 public class DetallePedidoDTO {
     private Integer id;
     private Integer idProducto;
     private Integer cantidad;
-    private Integer precioUnitario;
+    private BigDecimal precioUnitario;
 }
